extract select_chain module
